pyproject.toml
[tool.poetry]
name = "arhelp"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
frame-sdk = "^1.2.1"
openai = "^1.38.0"
pillow = "^10.4.0"
aiohttp = "^3.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
